Refactor craft type package
