+++ modules/Movecraft/src/main/java/net/countercraft/movecraft/processing/functions/DetectionPredicate.java
@@ -1,6 +1,6 @@
 package net.countercraft.movecraft.processing.functions;
 
-import net.countercraft.movecraft.craft.CraftType;
+import net.countercraft.movecraft.craft.type.CraftType;
 import net.countercraft.movecraft.processing.MovecraftWorld;
 import org.bukkit.command.CommandSender;
 import org.jetbrains.annotations.Contract;
