+++ modules/Movecraft/src/main/java/net/countercraft/movecraft/processing/tasks/detection/validators/SubcraftValidator.java
@@ -3,7 +3,7 @@ package net.countercraft.movecraft.processing.tasks.detection.validators;
 import net.countercraft.movecraft.MovecraftLocation;
 import net.countercraft.movecraft.craft.Craft;
 import net.countercraft.movecraft.craft.CraftManager;
-import net.countercraft.movecraft.craft.CraftType;
+import net.countercraft.movecraft.craft.type.CraftType;
 import net.countercraft.movecraft.localisation.I18nSupport;
 import net.countercraft.movecraft.processing.MovecraftWorld;
 import net.countercraft.movecraft.processing.functions.Result;
