+++ modules/Movecraft/src/main/java/net/countercraft/movecraft/processing/tasks/translation/validators/HoverValidator.java
@@ -1,12 +1,11 @@
 package net.countercraft.movecraft.processing.tasks.translation.validators;
 
 import net.countercraft.movecraft.MovecraftLocation;
-import net.countercraft.movecraft.craft.CraftType;
+import net.countercraft.movecraft.craft.type.CraftType;
 import net.countercraft.movecraft.localisation.I18nSupport;
 import net.countercraft.movecraft.processing.MovecraftWorld;
 import net.countercraft.movecraft.processing.functions.Result;
 import net.countercraft.movecraft.processing.functions.TetradicPredicate;
-import net.countercraft.movecraft.processing.functions.TriadicPredicate;
 import net.countercraft.movecraft.util.hitboxes.HitBox;
 import org.bukkit.Material;
 import org.jetbrains.annotations.Contract;
