+++ modules/Movecraft/src/main/java/net/countercraft/movecraft/processing/tasks/translation/validators/WorldBorderValidator.java
@@ -1,7 +1,7 @@
 package net.countercraft.movecraft.processing.tasks.translation.validators;
 
 import net.countercraft.movecraft.MovecraftLocation;
-import net.countercraft.movecraft.craft.CraftType;
+import net.countercraft.movecraft.craft.type.CraftType;
 import net.countercraft.movecraft.localisation.I18nSupport;
 import net.countercraft.movecraft.processing.MovecraftWorld;
 import net.countercraft.movecraft.processing.functions.Result;
