+++ modules/api/src/main/java/net/countercraft/movecraft/craft/type/CraftType.java
@@ -15,7 +15,7 @@
  *     along with Movecraft.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package net.countercraft.movecraft.craft;
+package net.countercraft.movecraft.craft.type;
 
 import net.countercraft.movecraft.processing.MovecraftWorld;
 import net.countercraft.movecraft.util.Pair;
